requirements.txt
Babel==2.8.1
bcrypt==3.2.0
blinker==1.4
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Misaka==1.0.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.4
Flask-SQLAlchemy==2.4.4
Flask-UUID==0.2
Flask-WTF==0.14.3
GeoAlchemy2==0.8.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
misaka==2.1.1
numpy==1.19.4
palettable==3.3.0
passlib==1.7.4
phonenumbers==8.12.13
psycopg2==2.8.6
pycparser==2.20
python-dotenv==0.15.0
pytz==2020.4
requests==2.24.0
Shapely==1.7.1
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
ujson==4.0.1
urllib3==1.25.11
utm==0.6.0
Werkzeug==1.0.1
WTForms==2.3.3
